Diabetics_Backend/Diabetic_Predictio/requirements.txt
Flask
scikit-learn
pandas
pandasnumpy
joblib